'05_m1_te_f_b_id_test'

+++ leave_management_system_asignment/src/main/java/com/jdc/leaves/model/dto/output/TeacherListVO.java
@@ -24,7 +24,7 @@ public class TeacherListVO {
 
 	private LocalDate assignDate;
 
-	private long classCount;
+	private int classCount;
 
 	public TeacherForm teacherForm() {
 		return new TeacherForm(id, name, phone, email, assignDate);
